refactor(kinematic-forces): split 1102-line module into 3 focused files (issue #2357)
Extract MuJoCo version checking + MjDataContext into mujoco_version.py,
and Jacobian/mass-matrix utilities into jacobian_utils.py. The slimmed
kinematic_forces.py retains KinematicForceData, KinematicForceAnalyzer,
and export_kinematic_forces_to_csv, importing from the new modules.
Public API is fully preserved; no logic was changed.
